add api layer and cleanup
