+++ frontend/app/src/entities/path-traversal/ui/queries/get-path-traversal.query.ts
@@ -1,19 +1,24 @@
-import { type UseQueryOptions, useQuery } from "@tanstack/react-query";
+import { queryOptions, type UseQueryOptions, useQuery } from "@tanstack/react-query";
 import { useAtomValue } from "jotai";
 
 import { datetimeAtom } from "@/shared/stores/time.atom";
 
 import { useCurrentBranch } from "@/entities/branches/ui/branches-provider";
+import { getPathTraversal } from "@/entities/path-traversal/domain/get-path-traversal";
+import type {
+  GetPathTraversalParams,
+  PathTraversalResponse,
+} from "@/entities/path-traversal/domain/path-traversal.types";
+import { pathTraversalQueryKeys } from "@/entities/path-traversal/ui/queries/path-traversal.query-keys";
 
-import {
-  type GetPathTraversalParams,
-  getPathTraversal,
-  type PathTraversalResponse,
-} from "./get-path-traversal";
-import { pathTraversalKeys } from "./path-traversal.query-keys";
+export function getPathTraversalQueryOptions(params: GetPathTraversalParams) {
+  return queryOptions({
+    queryKey: pathTraversalQueryKeys.traverse(params),
+    queryFn: () => getPathTraversal(params),
+  });
+}
 
 type UseGetPathTraversalParams = Omit<GetPathTraversalParams, "branchName" | "atDate">;
-
 type UseGetPathTraversalConfig = Partial<UseQueryOptions<PathTraversalResponse>>;
 
 export function useGetPathTraversal(
@@ -24,17 +29,11 @@ export function useGetPathTraversal(
   const timeMachineDate = useAtomValue(datetimeAtom);
 
   return useQuery({
-    queryKey: pathTraversalKeys.traverse({
+    ...getPathTraversalQueryOptions({
       branchName: currentBranch.name,
       atDate: timeMachineDate,
       ...params,
     }),
-    queryFn: () =>
-      getPathTraversal({
-        branchName: currentBranch.name,
-        atDate: timeMachineDate,
-        ...params,
-      }),
     ...config,
   });
 }
