+++ frontend/app/src/entities/path-traversal/ui/queries/path-traversal.query-keys.ts
@@ -1,0 +1,12 @@
+import type {
+  GetPathTraversalParams,
+  GetReachableObjectsParams,
+} from "@/entities/path-traversal/domain/path-traversal.types";
+
+export const pathTraversalQueryKeys = {
+  all: ["path-traversal"] as const,
+  traverse: (params: GetPathTraversalParams) =>
+    [...pathTraversalQueryKeys.all, "traverse", params] as const,
+  reachable: (params: GetReachableObjectsParams) =>
+    [...pathTraversalQueryKeys.all, "reachable", params] as const,
+} as const;
